+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### External Facing Changes
 - Users can now delete their own folders (specifically the run-id folder) after a globus download. (jaws-support:#238)
+- Added rc to tasks command (#1832)
 
 ### Internal Facing Changes